Adding support for Logback-access events
